~ | static_cast to answer subject

+++ M-06/ex00/ScalarConverter.cpp
@@ -15,7 +15,7 @@ void ScalarConverter::convert(const std::string &literal) {
 	bool isPseudoLiteral(false);
 
 	toInt = atoi(literal.c_str());
-	toFloat = atof(literal.c_str());
+	toFloat = static_cast<float>(atof(literal.c_str()));
 	toDouble = atof(literal.c_str());
 
 	if (std::isprint(toInt)) {
